avformat/subtitles: Use size_t for len
string length could theoretically be larger than int

(cherry picked from commit a633928d47057426a9c328da594407d1c7da8a5c)

+++ libavformat/subtitles.h
@@ -116,7 +116,7 @@ typedef struct {
  *              previous one instead of adding a new entry, 0 otherwise
  */
 AVPacket *ff_subtitles_queue_insert(FFDemuxSubtitlesQueue *q,
-                                    const uint8_t *event, int len, int merge);
+                                    const uint8_t *event, size_t len, int merge);
 
 /**
  * Set missing durations and sort subtitles by PTS, and then byte position.
